Early Intro to 9
Still has some refresh issues and clearly goes slower than it should. Fixes a couple long standing bugs and should work on migrating the LaserProject God Object into reasonable subclasses for the renderer tree/nodes and a standard project hub. Which would be highly important to keep the front end GUI elements apart from the rest. So CLI or other interfaces remain reasonable. Also, includes some basic keymapping.
